Task1 refactoring
